devide_response_rule
Separate 'ResponseRules' from response_manager.
